requirements.txt
# 核心依赖
ccxt>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0

# 回测框架
backtrader>=1.9.78

# 数据可视化
matplotlib>=3.7.0
plotly>=5.14.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0

# 异步处理
aiohttp>=3.8.0
asyncio>=3.4.3

# 工具库
requests>=2.31.0
python-dateutil>=2.8.2

# 监控和通知
flask>=2.3.0
dash>=2.11.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

